feat: create two separate layouts
