+++ app/app/layout.tsx
@@ -1,5 +1,3 @@
-import "react-loading-skeleton/dist/skeleton.css";
-
 import Navbar from "@/components/Navigation";
 import {
   DesktopSidebar,
@@ -7,32 +5,21 @@ import {
   SidebarPanels,
 } from "@/components/Sidebar";
 
-import "./globals.css";
-import Providers from "./providers";
+import AppProviders from "./providers";
 
-const RootLayout = ({ children }: { children: React.ReactNode }) => (
-  <html lang="en" className="scroll-smooth">
-    <head>
-      <title>DKMS</title>
-      <meta content="width=device-width, initial-scale=1" name="viewport" />
-      <meta name="description" content="Generated by create next app" />
-    </head>
-    {/* Suppress theme errors: https://github.com/pacocoursey/next-themes#with-app */}
-    <body suppressHydrationWarning>
-      <Providers>
-        <div className="md:flex flex-row min-h-screen">
-          <Navbar />
+const AppLayout = ({ children }: { children: React.ReactNode }) => (
+  <AppProviders>
+    <div className="md:flex flex-row min-h-screen">
+      <Navbar />
 
-          <main role="main" className="container mx-auto p-4 md:px-8 flex-grow">
-            <SidebarPanels />
-            <MobileSidebar>{children}</MobileSidebar>
-          </main>
-          {/* @ts-expect-error Server Component */}
-          <DesktopSidebar />
-        </div>
-      </Providers>
-    </body>
-  </html>
+      <main role="main" className="container mx-auto p-4 md:px-8 flex-grow">
+        <SidebarPanels />
+        <MobileSidebar>{children}</MobileSidebar>
+      </main>
+      {/* @ts-expect-error Server Component */}
+      <DesktopSidebar />
+    </div>
+  </AppProviders>
 );
 
-export default RootLayout;
+export default AppLayout;
