+++ app/layout.tsx
@@ -1,0 +1,20 @@
+import "react-loading-skeleton/dist/skeleton.css";
+
+import "./globals.css";
+import Providers from "./providers";
+
+const RootLayout = ({ children }: { children: React.ReactNode }) => (
+  <html lang="en" className="scroll-smooth">
+    <head>
+      <title>DKMS</title>
+      <meta content="width=device-width, initial-scale=1" name="viewport" />
+      <meta name="description" content="Generated by create next app" />
+    </head>
+    {/* Suppress theme errors: https://github.com/pacocoursey/next-themes#with-app */}
+    <body suppressHydrationWarning>
+      <Providers>{children}</Providers>
+    </body>
+  </html>
+);
+
+export default RootLayout;
