+++ cypress/support/component.ts
@@ -20,7 +20,7 @@
 import { mount } from "cypress/react18";
 import "react-loading-skeleton/dist/skeleton.css";
 
-import "@/app/app/globals.css";
+import "@/app/globals.css";
 
 // Import commands.js using ES2015 syntax:
 import "./commands";
